Solved reflective BC for ABM + no self.init_epoch
BC where checked twice, one correctly by the virtual env, but then also once more by the ABM_motion simulator. We delete this second check, as it is not needed.

Also deleted the self init_epoch so that the epochs are controlled by the outer loop. Done this to mimic their code. Also now we output a done = False / True